Tests adjustments
